Use boolean instead of a flag.  This flag field was not a problem because it only had one flag defined.  But the best practice for coding in a multithreading environment is to use booleans instead of a flag field.

+++ storage/falcon/Cache.cpp
@@ -288,7 +288,7 @@ Bdb* Cache::fetchPage(Dbb *dbb, int32 pageNumber, PageType pageType, LockType lo
 
 	// If buffer has moved out of the upper "fraction" of the LRU queue, move it back up
 	
-	if (bdb->age < bufferAge - upperFraction)
+	if (bdb->age < bufferAge - (uint64) upperFraction)
 		{
 		sync.lock (Exclusive);
 		moveToHead (bdb);
